Use new Csar to show content in Directory and FileControllers
Also correct the bridging between CsarId and CSARID.
Additionally deprecates a constructor for CsarImpl that assumes CsarId
knows where the Csar itself is stored. That assumption should not be
something we get used to.

In reference to incorrect assumptions a few FIXMEs and TODOs were
changed, as well as multiple methods Deprecated