Move `deployment.tests` to `deployment.checks`
This deals with maven trying to be a smartass and not building tests modules
unless we're running tests. That resulted in compilation problems when
trying to use the classes from the deployment.tests module.

Renaming to deployment.checks gets us away from the magic "tests" string
at the end of the module name, but keeps the semantic meaning intact