Integrate Planbuilder
We pull a switcheroo on planbuilder's CSARHandler to use a different Csar Repository base path than core
to enable keeping with the pre-existing workflow of export -> plan -> import (overwrite) in core.

For that purpose CsarService needs to allow different Base Paths. This commit also cleans up the
responsibilities of CsarId, which now is no longer aware of any Filesystem. Down the line
this can allow for distributed storage or even in-memory storage of Csars.