Replace references to IToscaEngineService in ManagementBus
This also moves a few methods dealing with container-internal model types into the separate ContainerEngine helper.
Additionally the XMLSerializer is updated to rely on URLs for accessing the schema definitions of the TOSCA model
to enable easier dealing with the fact that the model is stored inside the packaged jar.