Remove IFileAccessService and it's implementation
This replaces all invocations of the java.io.File based IFileAccessService calls with
java.nio.Path based APIs implemented in a new abstraction called FileSystem.

That abstraction should be injectable down the line, but currently is just a static helper.
In addition this eschews the apache.commons library for zip handling, because it does not
handle Paths in the included version